+++ core/src/main/scala/tbd/list/ListConf.scala
@@ -19,7 +19,7 @@ case class ListConf
     (file: String = "",
      partitions: Int = 8,
      partitionIndex: Int = 0,
-     partitionSize: Int = 0,
+     partitionSize: Long = 0,
      chunkSize: Int = 1,
      chunkSizer: Any => Int = _ => 1,
      sorted: Boolean = false)
